+++ examples/minimal/outputs.tf
@@ -12,7 +12,8 @@ output "tamr_key_pair_name" {
 }
 
 output "tamr_private_key" {
-  value = tls_private_key.tamr_ec2_private_key.private_key_pem
+  value     = tls_private_key.tamr_ec2_private_key.private_key_pem
+  sensitive = true
 }
 
 output "tamr_vm" {